refactor: OSマップ機能を拡張し全未実装ページのプレースホルダを追加
- /src/config/osFeatures.ts に設定ファイルを移動
- PlannedFeaturePlaceholderコンポーネントを作成
- 全31個の未実装ページにプレースホルダを配置
  - 経営OS: KPI辞書、事業別サマリー、アラートセンター
  - 文書・契約: 契約書管理、同意書管理、電子署名ログ
  - 人・権限: 権限ロール管理、組織ツリー、外部関係者アカウント
  - 周知: 周知事項、既読管理、申し送り、通知センター
  - 申請・稟議: 承認ログ、差戻し管理
  - 業務管理: チケット管理、修繕チケット、備品在庫管理
  - 教育: 研修管理、資格管理、委員会管理、議事録、研修実施報告
  - リスク: クレーム対応、是正措置管理
  - 利用者・家族: 家族連絡ログ、キーパーソン管理、連絡履歴
  - 財務: 未収管理、回収フロー、督促履歴
- 「存在したら表示」の思想を徹底
