+++ src/app/admin/external-accounts/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function ExternalAccountsPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="外部関係者アカウント"
+      description="外部関係者のアクセス管理。取引先、パートナー、監査人などの外部アクセスを管理します。"
+      category="人・権限"
+    />
+  );
+}
