+++ src/app/admin/org-tree/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function OrgTreePage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="組織ツリー"
+      description="組織構造の可視化。部署、チーム、レポートラインを視覚的に表示します。"
+      category="人・権限"
+    />
+  );
+}
