+++ src/app/admin/roles/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function RolesPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="権限ロール管理"
+      description="権限とロールの設定。ユーザーの権限レベル、アクセス可能な機能を設定します。"
+      category="人・権限"
+    />
+  );
+}
