+++ src/app/dashboard/alerts/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function AlertCenterPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="アラートセンター"
+      description="全アラートの一元管理。期限切れ、異常値、緊急対応が必要な項目を集約表示します。"
+      category="経営OS（Core）"
+    />
+  );
+}
