+++ src/app/dashboard/announcements/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function AnnouncementsPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="周知事項"
+      description="全体周知の管理。重要なお知らせ、ポリシー変更、イベント情報を共有します。"
+      category="周知・コミュニケーション"
+    />
+  );
+}
