+++ src/app/dashboard/approval-log/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function ApprovalLogPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="承認ログ"
+      description="承認履歴の管理。過去の承認・却下の記録を検索・確認できます。"
+      category="申請・稟議"
+    />
+  );
+}
