+++ src/app/dashboard/business-summary/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function BusinessSummaryPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="事業別サマリー"
+      description="事業ごとの業績サマリー。各事業の売上、利益、KPIを一覧で確認できます。"
+      category="経営OS（Core）"
+    />
+  );
+}
