+++ src/app/dashboard/certifications/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function CertificationsPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="資格管理"
+      description="資格の取得状況管理。従業員の資格、有効期限、更新予定を管理します。"
+      category="教育・ガバナンス"
+    />
+  );
+}
