+++ src/app/dashboard/collection-flow/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function CollectionFlowPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="回収フロー"
+      description="回収プロセスの管理。未収金の回収ステップと進捗を管理します。"
+      category="未収・財務補助"
+    />
+  );
+}
