+++ src/app/dashboard/committees/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function CommitteesPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="委員会管理"
+      description="委員会の運営管理。各委員会のメンバー、会議スケジュール、議事録を管理します。"
+      category="教育・ガバナンス"
+    />
+  );
+}
