+++ src/app/dashboard/complaints/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function ComplaintsPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="クレーム対応"
+      description="クレームの記録と対応。クレーム内容、対応状況、再発防止策を管理します。"
+      category="リスク・品質"
+    />
+  );
+}
