+++ src/app/dashboard/consent/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function ConsentPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="同意書管理"
+      description="同意書の取得状況管理。入居者・家族からの同意書取得状況を追跡します。"
+      category="文書・契約"
+    />
+  );
+}
