+++ src/app/dashboard/contact-history/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function ContactHistoryPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="連絡履歴"
+      description="全連絡の履歴管理。入居者、家族、取引先との全連絡を一元管理します。"
+      category="利用者・家族"
+    />
+  );
+}
