+++ src/app/dashboard/contracts/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function ContractsPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="契約書管理"
+      description="契約書の管理と期限アラート。契約更新日の通知、契約内容の確認ができます。"
+      category="文書・契約"
+    />
+  );
+}
