+++ src/app/dashboard/corrective-actions/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function CorrectiveActionsPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="是正措置管理"
+      description="是正措置の追跡。問題の根本原因分析と是正措置の実施状況を管理します。"
+      category="リスク・品質"
+    />
+  );
+}
