+++ src/app/dashboard/dunning-history/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function DunningHistoryPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="督促履歴"
+      description="督促の履歴管理。督促連絡の日時、内容、結果を記録します。"
+      category="未収・財務補助"
+    />
+  );
+}
