+++ src/app/dashboard/esign-log/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function EsignLogPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="電子署名ログ"
+      description="電子署名の履歴管理。署名日時、署名者、対象文書を記録・確認できます。"
+      category="文書・契約"
+    />
+  );
+}
