+++ src/app/dashboard/family-contact/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function FamilyContactPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="家族連絡ログ"
+      description="家族への連絡履歴。入居者の家族への連絡内容と日時を記録します。"
+      category="利用者・家族"
+    />
+  );
+}
