+++ src/app/dashboard/handover/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function HandoverPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="申し送り"
+      description="シフト間の申し送り。前シフトから次シフトへの重要事項を記録・共有します。"
+      category="周知・コミュニケーション"
+    />
+  );
+}
