+++ src/app/dashboard/inventory/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function InventoryPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="備品在庫管理"
+      description="備品の在庫管理。消耗品、設備の在庫状況と発注タイミングを管理します。"
+      category="業務管理"
+    />
+  );
+}
