+++ src/app/dashboard/key-person/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function KeyPersonPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="キーパーソン管理"
+      description="キーパーソンの情報管理。入居者の主要連絡先、関係性を管理します。"
+      category="利用者・家族"
+    />
+  );
+}
