+++ src/app/dashboard/kpi-dictionary/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function KpiDictionaryPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="KPI辞書"
+      description="KPI定義と計算ロジックの一覧。各KPIの意味、計算方法、目標値を確認できます。"
+      category="経営OS（Core）"
+    />
+  );
+}
