+++ src/app/dashboard/meeting-minutes/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function MeetingMinutesPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="議事録"
+      description="会議議事録の管理。会議内容、決定事項、アクションアイテムを記録します。"
+      category="教育・ガバナンス"
+    />
+  );
+}
