+++ src/app/dashboard/notifications/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function NotificationsPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="通知センター"
+      description="全通知の一元管理。システム通知、アラート、リマインダーを集約表示します。"
+      category="周知・コミュニケーション"
+    />
+  );
+}
