+++ src/app/dashboard/read-status/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function ReadStatusPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="既読管理"
+      description="既読状況の追跡。重要な周知事項やドキュメントの既読状況を確認します。"
+      category="周知・コミュニケーション"
+    />
+  );
+}
