+++ src/app/dashboard/receivables/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function ReceivablesPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="未収管理"
+      description="未収金の管理。未回収の請求、滞納状況を一覧で確認できます。"
+      category="未収・財務補助"
+    />
+  );
+}
