+++ src/app/dashboard/repair-tickets/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function RepairTicketsPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="修繕チケット"
+      description="修繕依頼の管理。設備故障、修理依頼の進捗を追跡します。"
+      category="業務管理"
+    />
+  );
+}
