+++ src/app/dashboard/returns/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function ReturnsPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="差戻し管理"
+      description="差戻し案件の管理。差戻しされた申請の対応状況を追跡します。"
+      category="申請・稟議"
+    />
+  );
+}
