+++ src/app/dashboard/tickets/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function TicketsPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="チケット管理"
+      description="業務チケットの管理。タスク、依頼、問い合わせをチケット形式で管理します。"
+      category="業務管理"
+    />
+  );
+}
