+++ src/app/dashboard/training-reports/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function TrainingReportsPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="研修実施報告"
+      description="研修実施の報告管理。研修の実施結果、参加者フィードバックを記録します。"
+      category="教育・ガバナンス"
+    />
+  );
+}
