+++ src/app/dashboard/training/page.tsx
@@ -1,0 +1,11 @@
+import { PlannedFeaturePlaceholder } from '@/components/PlannedFeaturePlaceholder';
+
+export default function TrainingPage() {
+  return (
+    <PlannedFeaturePlaceholder
+      title="研修管理"
+      description="研修の計画と実施管理。研修スケジュール、参加者、進捗を管理します。"
+      category="教育・ガバナンス"
+    />
+  );
+}
